Switched over to reusable, but kept custom details in new component
