+++ resources/js/pages/admin/tabs/ReportedUsers.vue
@@ -1,203 +1,73 @@
 <template>
     <div>
         <Loading v-if="loading" />
-        <div v-else class="reported-users-overview">
-            <h3>{{ $t('reported-users') }}</h3>
-            
-            <div class="header-row row">
-                <span class="col fg-1 header clickable" @click="setSort('username')">
-                    {{ $t('username') }}
-                    <FaIcon 
-                        icon="sort"  />
-                </span>
-                <span class="fg-1 col header clickable" @click="setSort('report_amount')">
-                    {{ $t('number-of-reports') }}
-                    <FaIcon 
-                        icon="sort"  />
-                </span>
-                <span class="col header clickable" @click="setSort('last_report_date')">
-                    {{ $t('last-report') }}
-                    <FaIcon 
-                        icon="sort"  />
-                </span>
-                <span class="col header clickable" @click="setSort('banned_until')">
-                    {{ $t('banned-until') }}
-                    <FaIcon 
-                        icon="sort"  />
-                </span>
-                <span class="fg-1 col header">
-                    {{ $t('actions') }}
-                </span>
-            </div>
 
-            <template v-for="(user, index) in sortedReportedUsers" :key="index">
-                <div class="detailed-table">
-                    <div class="row top-row">
-                        <div class="col fg-1">
-                            {{user.username}}
-                        </div>
-                        <div class="col fg-1">
-                            <FaIcon 
-                                class="expand-button"
-                                icon="caret-down" 
-                                @click="showReportDetails(index)" />
-                            {{user.report_amount}}
-                        </div>
-                        <div class="col">
-                            {{parseDateTime(user.last_report_date)}}
-                        </div>
-                        <div class="col">
-                            <FaIcon 
-                                v-if="user.banned"
-                                class="expand-button"
-                                icon="caret-down" 
-                                @click="showBannedDetails(index)" />
-                            {{user.banned ? user.banned[0].banned_until_time : 'Never'}}
-                        </div>
-                        <div class="fg-1 col">
-                            <Tooltip v-if="!currentlyBanned(user)" :text="$t('suspend-user')">
-                                <FaIcon 
-                                    icon="ban" 
-                                    class="icon red"
-                                    @click="suspendUser(user)" />                        
-                            </Tooltip>
-                            <Tooltip :text="$t('message-reported-user')">
-                                <FaIcon 
-                                    icon="envelope"
-                                    class="icon"
-                                    @click="sendMessageToReportedUser(user)" />
-                            </Tooltip>
-                        </div>
-                    </div>
-                    <!-- Banned overview -->
-                    <div 
-                        v-if="user.banned"
-                        :ref="el => { bannedDivs[index] = el as HTMLDivElement }" class="sub-details row" 
-                        :style="{'max-height': '0px', transition: 'max-height 2s'}">
-                        <h5>{{ $t('bans') }}</h5>
-                        <div class="row">
-                            <span class="col header">{{ $t('start-ban') }}</span>
-                            <span class="col header">{{ $t('end-ban') }}</span>
-                            <span class="col header">{{ $t('reason') }}</span>
-                            <span class="col fg-1 header">{{ $t('days') }}</span>
-                            <span class="col fg-1 header">{{ $t('banned-by') }}</span>
-                        </div>
-                        <template v-for="(banned, index) in user.banned" :key="index">
-                            <div class="row">
-                                <div class="col">
-                                    {{parseDateTime(banned.created_at)}} ({{banned.time_since}})
-                                </div>
-                                <div class="col">
-                                    {{banned.early_release ? 
-                                        parseDateTime(banned.early_release) : parseDateTime(banned.banned_until)}} 
-                                    ({{banned.banned_until_time}})
-                                </div>
-                                <div class="col">
-                                    {{banned.reason}}
-                                </div>
-                                <div class="col fg-1">
-                                    {{banned.days}}
-                                </div>
-                                <div class="col fg-1">
-                                    {{banned.admin.username}}
-                                </div>
-                            </div>
-                        </template>
-                    </div>
-                    <!-- Reports overview -->
-                    <div 
-                        :ref="el => { reportDivs[index] = el as HTMLDivElement }" class="sub-details row" 
-                        :style="{'max-height': '0px'}">
-                        <h5>{{ $t('reports') }}</h5>
-                        <div class="row">
-                            <span class="col fg-1 header">{{ $t('reason') }}</span>
-                            <span class="col header">{{ $t('comment') }}</span>
-                            <span class="col header">{{ $t('date') }}</span>
-                            <span class="col header">{{ $t('reported-by') }}</span>
-                            <span class="col fg-1 header">{{ $t('conversation') }}</span>
-                        </div>
-                        <template v-for="(report, index) in user.reports" :key="index">
-                            <div class="row">
-                                <div class="col fg-1">
-                                    {{parseReason(report.reason)}}
-                                </div>
-                                <div class="col">
-                                    {{report.comment}}
-                                </div>
-                                <div class="col">
-                                    {{parseDateTime(report.reported_date)}}
-                                </div>
-                                <div class="col">
-                                    {{report.reported_by_name}}
-                                </div>
-                                <div class="col fg-1">
-                                    <div v-if="report.conversation">
-                                        {{ $t('yes')}}
-                                        <Tooltip :text="$t('show-conversation')">
-                                            <FaIcon 
-                                                icon="magnifying-glass" 
-                                                class="icon"
-                                                @click="showConversation(report.conversation)" />
-                                        </Tooltip>
-                                    </div>
-                                    <div v-else>
-                                        {{ $t('no')}}
-                                    </div>
-                                </div>
-                            </div>
-                        </template>
-                    </div>
-                </div>
+        <SortableOverviewTable
+            v-if="sortedReportedUsers"
+            :items="sortedReportedUsers"
+            :fields="REPORTED_USER_OVERVIEW_FIELDS"
+            click-to-extend>
+
+            <template #username="item">
+                <h5>{{item.item.username}}</h5>
             </template>
-            <Modal :show="showConversationModal"
-                   :footer="false" 
-                   :title="$t('see-conversation-reporter')" 
-                   @close="closeShowConversation">
-                <ShowConversationModal 
-                    v-if="conversationIdToShow !== null" 
-                    :conversationId="parseInt(conversationIdToShow)" 
-                    @close="closeShowConversation"/>
-            </Modal>
-            <Modal
-                :show="showSuspendUserModal"
-                :footer="false"
-                :title="suspendUserTitle"
-                @close="closeSuspendUserModal">
-                <SuspendUserModal 
-                    v-if="suspendedUser !== null"
-                    :userId="suspendedUser.id" 
-                    @close="closeSuspendUserModal" />
-            </Modal>
-            <Modal 
-                :show="showSendMessageModal" 
-                :footer="false" 
-                :title="$t('send-message')" 
-                @close="closeSendMessageToReportedUser">
-                <SendMessage v-if="userToMessage !== null" :user="userToMessage" @close="closeSendMessageToReportedUser"/>
-            </Modal>
-        </div>
+            <template #banned_until="item">
+                {{item.item.banned ? item.item.banned[0].banned_until_time : 'Never'}}
+            </template>
+            <template #actions="item">
+                <Tooltip v-if="!currentlyBanned(item.item)" :text="$t('suspend-user')">
+                    <FaIcon 
+                        icon="ban" 
+                        class="icon red"
+                        @click.stop.prevent="suspendUser(item.item)" />                        
+                </Tooltip>
+                <Tooltip :text="$t('message-reported-user')">
+                    <FaIcon 
+                        icon="envelope"
+                        class="icon"
+                        @click.stop.prevent="sendMessageToReportedUser(item.item)" />
+                </Tooltip>
+            </template>
+            <template #extend="item">
+                <ReportedUserDetails :user="item.item" />
+            </template>
+        </SortableOverviewTable>
+
+        <Modal
+            :show="showSuspendUserModal"
+            :footer="false"
+            :title="suspendUserTitle"
+            @close="closeSuspendUserModal">
+            <SuspendUserModal 
+                v-if="suspendedUser !== null"
+                :userId="suspendedUser.id" 
+                @close="closeSuspendUserModal" />
+        </Modal>
+        <Modal 
+            :show="showSendMessageModal" 
+            :footer="false" 
+            :title="$t('send-message')" 
+            @close="closeSendMessageToReportedUser">
+            <SendMessage v-if="userToMessage !== null" :user="userToMessage" @close="closeSendMessageToReportedUser"/>
+        </Modal>
     </div>
 </template>
 
 <script setup lang="ts">
 /* eslint-disable max-lines */
-import {ref, computed, onMounted, onBeforeUpdate} from 'vue';
+import {ref, computed, onMounted} from 'vue';
 import SuspendUserModal from './../components/SuspendUserModal.vue';
-import ShowConversationModal from '../components/ShowConversationModal.vue';
 import SendMessage from '/js/pages/messages/components/SendMessage.vue';
-import {parseDateTime} from '/js/services/dateService';
+import SortableOverviewTable from '/js/components/global/SortableOverviewTable.vue';
+import ReportedUserDetails from '../components/ReportedUsersDetails.vue';
 import {sortValues} from '/js/services/sortService';
 import {DateTime} from 'luxon';
 import {useAdminStore} from '/js/store/adminStore';
 import {useMainStore} from '/js/store/store';
-import {User, UserToBan} from 'resources/types/user';
+import {REPORTED_USER_OVERVIEW_FIELDS} from '/js/constants/reportUserConstants';
+import type {User, UserToBan} from 'resources/types/user';
 const adminStore = useAdminStore();
 const mainStore = useMainStore();
-
-onBeforeUpdate(() => {
-    reportDivs.value = [];
-    bannedDivs.value = [];
-});
 
 onMounted(async() => {
     await adminStore.getReportedUsers();
@@ -214,32 +84,6 @@ const sortedReportedUsers = computed(() => {
     if (reportedUsers.value === null) return;
     return sortValues(reportedUsers.value, currentSort.value, currentSortDir.value);
 });
-
-function setSort(key: string) {
-    if (currentSort.value == key) toggleSortDir();
-    else currentSort.value = key;
-}
-function toggleSortDir() {
-    if (currentSortDir.value == 'asc') currentSortDir.value = 'desc';
-    else currentSortDir.value = 'asc';
-}
-
-function parseReason(reason: string) {
-    let string = reason.replaceAll('_', ' ').toLowerCase();
-    return string.charAt(0).toUpperCase() + string.slice(1);
-}
-/** Opens the modal to see a conversation between reported user and reporter */
-const conversationIdToShow = ref<string | null>(null);
-const showConversationModal = ref(false);
-function showConversation(conversationId: string) {
-    mainStore.clearErrors();
-    conversationIdToShow.value = conversationId;
-    showConversationModal.value = true;
-}
-function closeShowConversation() {
-    conversationIdToShow.value = null;
-    showConversationModal.value = false;
-}
 
 /** Opens the modal to message a user that has been reported */
 const userToMessage = ref<User | null>(null);
@@ -270,17 +114,6 @@ function closeSuspendUserModal() {
     showSuspendUserModal.value = false;
 }
 
-const reportDivs = ref<Array<HTMLDivElement>>([]);
-const bannedDivs = ref<Array<HTMLDivElement>>([]);
-
-function showReportDetails(index: number) {
-    if (reportDivs.value[index].style.maxHeight == '0px') reportDivs.value[index].style.maxHeight = 'max-content';
-    else reportDivs.value[index].style.maxHeight = '0px';
-}
-function showBannedDetails(index: number) {
-    if (bannedDivs.value[index].style.maxHeight == '0px') bannedDivs.value[index].style.maxHeight = 'max-content';
-    else bannedDivs.value[index].style.maxHeight = '0px';
-}
 function currentlyBanned(user: UserToBan) {
     return !!user.banned_until && DateTime.now() < DateTime.fromFormat(user.banned_until, 'yyyy-MM-dd HH:mm:ss')
 }
